minor tweaks.
Here were some minor tweaks to meet certain coding expectations.
paragraph elements were added. As was the utf-8 character set in the
head. Additionally, we added a viewport specification.  To-do: add
srcset to images along with alts. Customize site with nav bar and/or
polymer elements.

In the CSS we changed several things to meet coding stylistic
conventions. Lowercase hex code for color, change of the logo ID to a
class, the removal of the border-style line for the hr element.